requirements.txt
certifi==2025.1.31
dnspython==2.7.0
eth-account==0.13.5
Flask==3.1.0
flask-cors==5.0.1
Flask-Limiter==3.10.1
Flask-Session==0.8.0
gunicorn==23.0.0
Jinja2==3.1.5
limits==4.0.1
markdown2==2.5.3
openai==1.64.0
pymongo==4.11.1
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.3.0
web3==7.8.0
Werkzeug==3.1.3
gunicorn==23.0.0